requirements.txt
streamlit
langchain
google-generativeai
sentence-transformers
faiss-cpu
pypdf
python-docx
unstructured
pytesseract